+++ src/main/java/com/revolut/transferservice/service/AccountService.java
@@ -1,5 +1,7 @@
 package com.revolut.transferservice.service;
 
+import java.math.BigDecimal;
+
 public interface AccountService {
-    void transfer(long idFrom, long idTo, double amount);
+    void transfer(long idFrom, long idTo, BigDecimal amount);
 }
